scheduler/requirements.txt
requests
pyyaml
psycopg2
pandas
sqlalchemy
joblib
scikit-learn